fix: change incorrect id type

+++ src/main/java/com/mlastovsky/repository/DirectorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DirectorRepository extends CrudRepository<Director, Integer> {
+public interface DirectorRepository extends CrudRepository<Director, Long> {
 
 }
